Some corrections for compilation.

+++ src_group/WeightMatrix.h
@@ -1457,10 +1457,10 @@ bool RenormalizeWeightMatrix(
     if (ListWeight1[iEnt] != ListWeight2[iEnt]) {
       std::cerr << "iEnt=" << iEnt << " weight1=" << ListWeight1[iEnt] << " weight2=" << ListWeight2[iEnt] << "\n";
       std::cerr << "Full weights=\n";
-      for (int u=0; u<nbEnt; u++) {
+      for (size_t u=0; u<nbEnt; u++) {
         std::cerr << "u=" << u << " weight1=" << ListWeight1[u] << " weight2=" << ListWeight2[u] << "\n";
       }
-      for (int u=0; u<nbEnt; u++) {
+      for (size_t u=0; u<nbEnt; u++) {
         std::cerr << "u=" << u << " gListRev=" << gListRev[u] << "\n";
       }
       std::cerr << "ERROR: The reordering failed\n";
